Regrouped user modal setting sections
Ref https://linear.app/ghost/issue/PLG-403/implement-tabs-in-user-modal
- Moved around the settings to be displayed in their new respective sections: Profile, Social links, and Email notifications.
