+++ apps/admin-x-settings/src/components/settings/general/users/RoleSelector.tsx
@@ -10,13 +10,13 @@ const RoleSelector: React.FC<{ user: User; setUserData: (user: User) => void; }>
 
     if (isOwnerUser(user)) {
         return (
-            <>
-                <Heading level={6}>Role</Heading>
-                <div className='flex flex-col items-center justify-center gap-3 bg-grey-75 px-10 py-12 text-center text-sm text-grey-800 dark:bg-grey-950 dark:text-white'>
+            <div>
+                <Heading useLabelTag={true}>Role</Heading>
+                <div className='mt-1.5 flex flex-col items-center justify-center gap-3 rounded-lg bg-grey-75 px-8 py-10 text-center text-sm text-grey-800 dark:bg-grey-950 dark:text-white'>
                     <Icon colorClass='text-grey-800 dark:text-white' name='crown' size='lg' />
                     This user is the owner of the site. To change their role, you need to transfer the ownership first.
                 </div>
-            </>
+            </div>
         );
     }
     let optionsArray = [
