+++ apps/admin-x-settings/src/components/settings/general/users/SocialLinksTab.tsx
@@ -1,5 +1,5 @@
 import CustomHeader from './CustomHeader';
-import {SettingGroup, SettingGroupContent, TextArea, TextField} from '@tryghost/admin-x-design-system';
+import {SettingGroup, SettingGroupContent, TextField} from '@tryghost/admin-x-design-system';
 import {UserDetailProps} from '../UserDetailModal';
 import {facebookHandleToUrl, facebookUrlToHandle, twitterHandleToUrl, twitterUrlToHandle, validateFacebookUrl, validateTwitterUrl} from '../../../../utils/socialUrls';
 
@@ -11,16 +11,6 @@ export const DetailsInputs: React.FC<UserDetailProps> = ({errors, clearError, va
 
     return (
         <SettingGroupContent>
-            <TextField
-                error={!!errors?.location}
-                hint={errors?.location || 'Where in the world do you live?'}
-                maxLength={65535}
-                title="Location"
-                value={user.location || ''}
-                onChange={(e) => {
-                    setUserData({...user, location: e.target.value});
-                }}
-                onKeyDown={() => clearError('location')} />
             <TextField
                 error={!!errors?.website}
                 hint={errors?.website || 'Have a website or blog other than this one? Link it!'}
@@ -35,8 +25,25 @@ export const DetailsInputs: React.FC<UserDetailProps> = ({errors, clearError, va
                 }}
                 onKeyDown={() => clearError('url')} />
             <TextField
+                error={!!errors?.twitter}
+                hint={errors?.twitter}
+                maxLength={2000}
+                title="X profile"
+                value={twitterUrl}
+                onBlur={(e) => {
+                    if (validateField('twitter', e.target.value)) {
+                        const url = validateTwitterUrl(e.target.value);
+                        setTwitterUrl(url);
+                        setUserData({...user, twitter: twitterUrlToHandle(url)});
+                    }
+                }}
+                onChange={(e) => {
+                    setTwitterUrl(e.target.value);
+                }}
+                onKeyDown={() => clearError('twitter')} />
+            <TextField
                 error={!!errors?.facebook}
-                hint={errors?.facebook || 'URL of your personal Facebook Profile'}
+                hint={errors?.facebook}
                 maxLength={2000}
                 title="Facebook profile"
                 value={facebookUrl}
@@ -51,42 +58,15 @@ export const DetailsInputs: React.FC<UserDetailProps> = ({errors, clearError, va
                     setFacebookUrl(e.target.value);
                 }}
                 onKeyDown={() => clearError('facebook')} />
-            <TextField
-                error={!!errors?.twitter}
-                hint={errors?.twitter || 'URL of your X profile'}
-                maxLength={2000}
-                title="X (formerly Twitter) profile"
-                value={twitterUrl}
-                onBlur={(e) => {
-                    if (validateField('twitter', e.target.value)) {
-                        const url = validateTwitterUrl(e.target.value);
-                        setTwitterUrl(url);
-                        setUserData({...user, twitter: twitterUrlToHandle(url)});
-                    }
-                }}
-                onChange={(e) => {
-                    setTwitterUrl(e.target.value);
-                }}
-                onKeyDown={() => clearError('twitter')} />
-            <TextArea
-                error={!!errors?.bio}
-                hint={errors?.bio || <>Recommended: 200 characters. You&lsquo;ve used <span className='font-bold'>{user.bio?.length || 0}</span></>}
-                maxLength={65535}
-                title="Bio"
-                value={user.bio || ''}
-                onChange={(e) => {
-                    setUserData({...user, bio: e.target.value});
-                }}
-                onKeyDown={() => clearError('bio')} />
         </SettingGroupContent>
     );
 };
 
-const ProfileDetails: React.FC<UserDetailProps> = (props) => {
+const SocialLinksTab: React.FC<UserDetailProps> = (props) => {
     return (
         <SettingGroup
             border={false}
-            customHeader={<CustomHeader>Details</CustomHeader>}
+            customHeader={<CustomHeader>Social links</CustomHeader>}
             title='Details'
         >
             <DetailsInputs {...props} />
@@ -94,4 +74,4 @@ const ProfileDetails: React.FC<UserDetailProps> = (props) => {
     );
 };
 
-export default ProfileDetails;
+export default SocialLinksTab;
